separate VAE blocks & VAE layers
